Storage: Move common functions into storage_common.c
