+++ src/CMakeLists.txt
@@ -11,6 +11,7 @@ target_sources_ifdef(CONFIG_THINGSET_LOG_BACKEND app PRIVATE log_backend.c)
 target_sources_ifdef(CONFIG_THINGSET_LORAWAN app PRIVATE lorawan.c)
 target_sources_ifdef(CONFIG_THINGSET_SERIAL app PRIVATE serial.c)
 target_sources_ifdef(CONFIG_THINGSET_SHELL app PRIVATE shell.c)
+target_sources_ifdef(CONFIG_THINGSET_STORAGE app PRIVATE storage_common.c)
 target_sources_ifdef(CONFIG_THINGSET_STORAGE_EEPROM app PRIVATE storage_eeprom.c)
 target_sources_ifdef(CONFIG_THINGSET_STORAGE_FLASH app PRIVATE storage_flash.c)
 target_sources_ifdef(CONFIG_THINGSET_WEBSOCKET app PRIVATE websocket.c)
