+++ src/storage_eeprom.c
@@ -38,8 +38,6 @@ struct thingset_eeprom_header
     uint32_t crc;
 } __packed;
 
-static struct k_work_delayable storage_work;
-
 static const struct device *eeprom_dev = DEVICE_DT_GET(EEPROM_DEVICE_NODE);
 
 static int thingset_eeprom_load(off_t offset)
@@ -288,32 +286,3 @@ int thingset_storage_save()
     return thingset_eeprom_save(0, eeprom_size);
 #endif
 }
-
-void thingset_storage_save_queued()
-{
-    thingset_sdk_reschedule_work(&storage_work, K_NO_WAIT);
-}
-
-static void regular_storage_handler(struct k_work *work)
-{
-    struct k_work_delayable *dwork = k_work_delayable_from_work(work);
-
-    thingset_storage_save();
-
-    if (IS_ENABLED(CONFIG_THINGSET_STORAGE_REGULAR)) {
-        thingset_sdk_reschedule_work(dwork, K_HOURS(CONFIG_THINGSET_STORAGE_INTERVAL));
-    }
-}
-
-static int thingset_storage_init(void)
-{
-    k_work_init_delayable(&storage_work, regular_storage_handler);
-
-    if (IS_ENABLED(CONFIG_THINGSET_STORAGE_REGULAR)) {
-        thingset_sdk_reschedule_work(&storage_work, K_HOURS(CONFIG_THINGSET_STORAGE_INTERVAL));
-    }
-
-    return 0;
-}
-
-SYS_INIT(thingset_storage_init, APPLICATION, THINGSET_INIT_PRIORITY_STORAGE);
